# tools-workspace/contracts/pdf/openapi.yaml
...
openapi: 3.0.3
info:
  title: EasyToolHub Tool API — PDF domain
  version: 0.1.0
  description: Client-upload PDF microservice for EasyToolHub (temp processing, no permanent storage).
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/pdf/preview/session:
    post:
      summary: Open an ephemeral page-preview session
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                password: { type: string, description: Base64-encoded UTF-8 password }
      responses:
        "200":
          description: Session metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId: { type: string }
                  pageCount: { type: integer }
                  maxDpi: { type: integer }
  /api/v1/pdf/preview/{sessionId}/pages/{page}.png:
    get:
      summary: Render one preview page as PNG
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
        - in: path
          name: page
          required: true
          schema: { type: integer, minimum: 1 }
        - in: query
          name: dpi
          schema: { type: integer, default: 120 }
      responses:
        "200":
          description: PNG
          content:
            image/png:
              schema: { type: string, format: binary }
...